APPLIED changed from revision 1
