+++ src/scripts/run_generic_problem_kronecker_graphs.sh
@@ -5,15 +5,16 @@ export RESULTS_DIR=${3}
 THREADS=( 32 )
 
 if [ "${problem}" = "tc" ]; then
-  export BINARIES=(${problem}_base ${problem}_bf ${problem}_1h ${problem}_doulion ${problem}_colorful )
+	export BINARIES=(${problem}_base ${problem}_bf ${problem}_1h ${problem}_doulion ${problem}_colorful ${problem}_pgp ${problem}_redex)
 else
-  export BINARIES=(${problem}_base ${problem}_bf ${problem}_1h)
+  export BINARIES=(${problem}_base ${problem}_bf ${problem}_1h ${problem}_pgp ${problem}_redex)
 fi
 
-PARAM_THs=(0.00003 0.0001 0.0003 0.001 0.003 0.01 0.03 0.1 0.3 0.5 1.0 2.0)
-
-KRON_SIZEs=(18 20 22)
-KRON_EDGEs=(1024 4096)
+REDEX_THs=(0.5)
+BF_THs=(0.5)
+OH_THs=(0.01)
+KRON_SIZEs=(16 18 20 22)
+KRON_EDGEs=(256 1024 4096)
 CLUSTER=0.1
 
 function create_launch {
@@ -111,24 +112,34 @@ for T in ${THREADS[@]}; do
         approx_scheme=$(echo ${BINARY} | cut -d'_' -f 2)
         
         if [ "${approx_scheme}" = "base" ]; then
-          export ARGS="-s -g ${KRON_SIZE} -k ${KRON_EDGE} -a -n 1 -y ${CLUSTER}"
-          create_launch_baseline
+          	export ARGS="-s -g ${KRON_SIZE} -k ${KRON_EDGE} -a -n 1 -y ${CLUSTER}"
+          	create_launch_baseline
         elif [ "${approx_scheme}" = "colorful" ]; then
-          export ARGS="-s -g ${KRON_SIZE} -k ${KRON_EDGE} -a -n 1 -p 0.5"
-          create_launch_baseline
+          	export ARGS="-s -g ${KRON_SIZE} -k ${KRON_EDGE} -a -n 1 -p 0.5"
+          	create_launch_baseline
         elif [ "${approx_scheme}" = "doulion" ]; then
 	        export ARGS="-s -g ${KRON_SIZE} -k ${KRON_EDGE} -a -n 1 -p 0.8"
-          create_launch_baseline
+          	create_launch_baseline
+	elif [ "${approx_scheme}" = "pgp" ] || [ "${approx_scheme}" = "redex" ]; then
+		for PARAM_TH in ${REDEX_THs[@]}; do
+			export ARGS="-s -g ${KRON_SIZE} -k ${KRON_EDGE} -a -n 1 -t ${PARAM_TH} -y ${CLUSTER}"
+			create_launch
+		done
+	elif [ "${approx_scheme}" = "approx1" ] || [ "${approx_scheme}" = "approx2" ]; then
+		for PARAM_TH in ${AUTO_THs[@]}; do
+			export ARGS="-s -g ${KRON_SIZE} -k ${KRON_EDGE} -a -n 1 -t ${PARAM_TH} -y ${CLUSTER}"
+			create_launch
+		done
         elif [ "${approx_scheme}" = "1h" ]; then
-	        for PARAM_TH in ${PARAM_THs[@]};
-            export ARGS="-s -g ${KRON_SIZE} -k ${KRON_EDGE} -t ${PARAM_TH} -n 1 -y ${CLUSTER}"
-            create_launch
-          done
-        else
-	        for PARAM_TH in ${PARAM_THs[@]};
-            export ARGS="-s -g ${KRON_SIZE} -k ${KRON_EDGE} -t ${PARAM_TH} -b -1 -n 1 -y ${CLUSTER}"
-            create_launch
-          done          
+	  	for PARAM_TH in ${OH_THs[@]}; do
+            		export ARGS="-s -g ${KRON_SIZE} -k ${KRON_EDGE} -t ${PARAM_TH} -n 1 -y ${CLUSTER}"
+            		create_launch
+    	  	done
+	elif [ "${approx_scheme}" = "bf" ]; then
+		for PARAM_TH in ${BF_THs[@]}; do 
+            		export ARGS="-s -g ${KRON_SIZE} -k ${KRON_EDGE} -t ${PARAM_TH} -b -1 -n 1 -y ${CLUSTER}"
+			create_launch
+		done          
         fi
       done
     done
